% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcdfunctions.R
\name{cart2genbary}
\alias{cart2genbary}
\title{Given the Cartesian coordinates of a set of points,
compute the generalized barycentric coordinates of these
points with respect to an outer simplex using the method of Warren 1996.}
\usage{
cart2genbary(Osimp, x)
}
\arguments{
\item{Osimp}{A 2d-by-d matrix representing a outer simplex}

\item{x}{A set of points of the target class}
}
\value{
The barycentric coordinates of the each points of x.
}
\description{
Given the Cartesian coordinates of a set of points,
compute the generalized barycentric coordinates of these
points with respect to an outer simplex using the method of Warren 1996.
}
\references{
Joe Warren. Barycentric Coordinates for Convex Polytopes. Advances in Computational Mathematics, 6:97–108, 1996.
}
