% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcdfunctions.R
\name{dist_to_plane}
\alias{dist_to_plane}
\title{Calculate the distance from a set of points to a plane}
\usage{
dist_to_plane(p, plane)
}
\arguments{
\item{p}{m-by-d matrix in which each row is the Cartesian coordinates of a point.}

\item{plane}{d-by-d matrix representing a plane.}
}
\value{
The set of distances between each point in p and a plane
}
\description{
Calculate the distance from a set of points to a plane
}
