% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dominate_brute_matrix}
\alias{dominate_brute_matrix}
\title{Finding the exact minimum dominating set of graph given its adjacency matrix
with the brute force algorithm (checking all possible solutions)}
\usage{
dominate_brute_matrix(A)
}
\arguments{
\item{A}{The adjacency matrix}
}
\value{
S A subset of the vertex set of the graph given the adjacency matrix A
}
\description{
Finding the exact minimum dominating set of graph given its adjacency matrix
with the brute force algorithm (checking all possible solutions)
}
