% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dominate_greedy_matrix}
\alias{dominate_greedy_matrix}
\title{Finding the approximately minimum dominating set of graph given its adjacency matrix
with the greedy algorithm}
\usage{
dominate_greedy_matrix(A)
}
\arguments{
\item{A}{The adjacency matrix}
}
\value{
S A subset (the approximate minimum dominating set) of the vertex set of the graph given the adjacency matrix A
}
\description{
Finding the approximately minimum dominating set of graph given its adjacency matrix
with the greedy algorithm
}
