% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{draw_outer_triangle}
\alias{draw_outer_triangle}
\title{Drawing a outer triangle (outer simplex in 2 dimensions)}
\usage{
draw_outer_triangle(outer_tri_list, ...)
}
\arguments{
\item{outer_tri_list}{A list of outer triangles, or a single outer triangle}

\item{...}{Additional arguements passed on to the polygon function}
}
\description{
Drawing a outer triangle (outer simplex in 2 dimensions)
}
