% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcdfunctions.R
\name{get_incenter_Osimp}
\alias{get_incenter_Osimp}
\title{Get the incenter of an outer simplex}
\usage{
get_incenter_Osimp(pla, m)
}
\arguments{
\item{pla}{N-by-N matrix representing a facet of the convex hull of the non-target class}

\item{m}{The median point of the convex hull of the non-target class}
}
\value{
The incenter of the outer simplex
}
\description{
Get the incenter of an outer simplex
}
