% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcdfunctions.R
\name{get_line_par_to}
\alias{get_line_par_to}
\title{Find a parallel ray passing the point, by means of orthogonal projection}
\usage{
get_line_par_to(x, ray)
}
\arguments{
\item{x}{The Cartesian coordinates of a reference point}

\item{ray}{2-by-N matrix representing a ray.}
}
\value{
A plane passing the point x which the parallel to the plane pla.
}
\description{
Find a parallel ray passing the point, by means of orthogonal projection
}
