% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcdfunctions.R
\name{get_line_plane_inter}
\alias{get_line_plane_inter}
\title{Find the point where the ray and the plane intersects}
\usage{
get_line_plane_inter(ray, pla)
}
\arguments{
\item{ray}{2-by-N matrix representing a ray.}

\item{pla}{N-by-N matrix representing a plane.}
}
\value{
A plane passing the point x which the parallel to the plane pla.
}
\description{
Find the point where the ray and the plane intersects
}
