% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcdfunctions.R
\name{get_plane_par_to}
\alias{get_plane_par_to}
\title{Find a parallel plane passing a point, by means of orthogonal projection}
\usage{
get_plane_par_to(x, pla)
}
\arguments{
\item{x}{The Cartesian coordinates of a reference point}

\item{pla}{N-by-N matrix representing a plane.}
}
\value{
A plane passing the point x which the parallel to the plane pla.
}
\description{
Find a parallel plane passing a point, by means of orthogonal projection
}
