% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcdfunctions.R
\name{getDisttoOsimp}
\alias{getDisttoOsimp}
\title{Get the distances between a set of points and a outer simplex.}
\usage{
getDisttoOsimp(Osimp, x)
}
\arguments{
\item{Osimp}{A 2d-by-d matrix representing a outer simplex.}

\item{x}{The set of points of the target class}
}
\value{
The set of distances between each point of x and a outer simplex
}
\description{
Get the distances between a set of points and a outer simplex.
}
