% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcdfunctions.R
\name{in_simp}
\alias{in_simp}
\title{Find if points are in a set of d-simplices}
\usage{
in_simp(simp, x)
}
\arguments{
\item{simp}{A list of d+1-by-d matrices each representing a d-simplex.}

\item{x}{The Cartesian coordinates of a reference point.}
}
\value{
A logical vector indicating in which simplex the reference point x resides.
}
\description{
Find if points are in a set of d-simplices
}
