% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcdfunctions.R
\name{outer_delaunayn}
\alias{outer_delaunayn}
\title{The list of the outer simplices of the Delaunay tessellation of the non-target class}
\usage{
outer_delaunayn(y)
}
\arguments{
\item{y}{The set of points of the non-target class}
}
\value{
The list of outer simplices of the Delaunay tessellation
}
\description{
The list of the outer simplices of the Delaunay tessellation of the non-target class
}
