% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcd.R
\name{pcd_cccd_multiclass_classify}
\alias{pcd_cccd_multiclass_classify}
\title{Classify a set of unlabelled data based on a PCD and a CCCD.}
\usage{
pcd_cccd_multiclass_classify(datan, dvalid, graph_pcd, p_pcd)
}
\arguments{
\item{datan}{The unlabelled data set}

\item{dvalid}{The distance matrix from the dominating set of CCCD to points in datan}

\item{graph_pcd}{A proximity catch digraph (PCD)}

\item{p_pcd}{The value of the parameter of the associated proximity map of the PCD}
}
\value{
Predicted labels of the unlabelled data set.
}
\description{
Classify a set of unlabelled data based on a PCD and a CCCD.
}
