% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcd-cs.R
\name{pcd_cs_Osimp}
\alias{pcd_cs_Osimp}
\title{Find the central similarity PCD (CS-PCD) of the target class restricted to a single outer simplex for a given parameter tau.}
\usage{
pcd_cs_Osimp(x, y, CM, tau)
}
\arguments{
\item{x}{The set of points of the target class}

\item{y}{The coordinates of the vertices of the  outer simplex}

\item{CM}{The median point of the convex hull of the non-target class}

\item{tau}{The CS-PCD parameter}
}
\value{
A list of objects associated with the dominating set of the CS-PCD restricted to a outer simplex
}
\description{
Find the central similarity PCD (CS-PCD) of the target class restricted to a single outer simplex for a given parameter tau.
}
