% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcd.R
\name{pcd_inhull_classifier}
\alias{pcd_inhull_classifier}
\title{The PCD classifier based on the proximity regions of the target class which are restricted to the inside of the convex hull of
the non-target class}
\usage{
pcd_inhull_classifier(datax, classes, method = "pe", param = 1)
}
\arguments{
\item{datax}{The labelled data set}

\item{classes}{A categorical variable indicating the classes of the points in the data set}

\item{method}{The Proximity Map associated with the classifier. "pe" for the Proportional-edge proximity maps
and "cs" for the Central-similarity proximity maps.}

\item{param}{The value of the parameter associated with the proximity map.}
}
\value{
A proximity catch digraph with the minimum dominating set and the associated proximity regions
}
\description{
The PCD classifier based on the proximity regions of the target class which are restricted to the inside of the convex hull of
the non-target class
}
