% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcd.R
\name{pcd_inhull_classify}
\alias{pcd_inhull_classify}
\title{Classify a set of unlabelled data based on a PCD. Returns NA for points outside of the convex hull of the non-target classes.}
\usage{
pcd_inhull_classify(datan, graph, param)
}
\arguments{
\item{datan}{The unlabelled data set}

\item{graph}{A proximity catch digraph (PCD)}

\item{param}{The value of the parameter of the associated proximity map of the PCD}
}
\value{
Predicted labels of the unlabelled data set.
}
\description{
Classify a set of unlabelled data based on a PCD. Returns NA for points outside of the convex hull of the non-target classes.
}
