% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcd.R
\name{pcd_multiclass_classifier}
\alias{pcd_multiclass_classifier}
\title{The PCD classifier based on the proximity regions of the target class where points inside the
convex hull are classified with n-simplicies, and outside with outer simplicies}
\usage{
pcd_multiclass_classifier(datax, classes, method = "pe", p_pcd = 1)
}
\arguments{
\item{datax}{The labelled data set}

\item{classes}{A categorical variable indicating the classes of the points in the data set}

\item{method}{The Proximity Map associated with the classifier. "pe" for the Proportional-edge proximity maps
and "cs" for the Central-similarity proximity maps.}

\item{p_pcd}{A set of values of the parameter associated with the proximity map of the PCD.}
}
\value{
A proximity catch digraph with the minimum dominating set and
        the associated proximity regions.
}
\description{
The PCD classifier based on the proximity regions of the target class where points inside the
convex hull are classified with n-simplicies, and outside with outer simplicies
}
