% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcd.R
\name{pcd_multiclass_classify}
\alias{pcd_multiclass_classify}
\title{Classify a set of unlabelled data based on a PCD.}
\usage{
pcd_multiclass_classify(datan, graph_pcd, p)
}
\arguments{
\item{datan}{The unlabelled data set}

\item{graph_pcd}{A proximity catch digraph (PCD)}

\item{p}{The value of the parameter of the associated proximity map of the PCD}
}
\value{
Predicted labels of the unlabelled data set.
}
\description{
Classify a set of unlabelled data based on a PCD.
}
