% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcd-pe.R
\name{pcd_pe_Osimp}
\alias{pcd_pe_Osimp}
\title{Find the pcd of an outer simplex for a given parameter r.}
\usage{
pcd_pe_Osimp(x, y, r)
}
\arguments{
\item{x}{The set of points of the target class}

\item{y}{The coordinates of the vertices of the outer simplex}

\item{r}{The PE-PCD parameter}
}
\value{
The dominating point of the outer simplex and the associated proximity region
}
\description{
Find the pcd of an outer simplex for a given parameter r.
}
