% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcdfunctions.R
\name{cart2genbary}
\alias{cart2genbary}
\title{Given the Cartesian coordinates of one or more points, compute the generalized barycentric coordinates of these
points with respect to an outer simplex using the method of Warren 1996.}
\usage{
cart2genbary(Osimp, x)
}
\arguments{
\item{Osimp}{the outer}

\item{x}{A set of points of the target class}
}
\value{
The barycentric coordinates of the each points of x.
}
\description{
Given the Cartesian coordinates of one or more points, compute the generalized barycentric coordinates of these
points with respect to an outer simplex using the method of Warren 1996.
}
