% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{draw_triangle}
\alias{draw_triangle}
\title{Drawing a triangle}
\usage{
draw_triangle(tri_list, ...)
}
\arguments{
\item{tri_list}{A list of triangles, or a single triangle}

\item{...}{Additional arguements passed on to the polygon function}
}
\description{
Drawing a triangle
}
