% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcdfunctions.R
\name{etsearchn}
\alias{etsearchn}
\title{Search for the outer simplicies that a set of points belong to}
\usage{
etsearchn(y, Osimp, x)
}
\arguments{
\item{y}{The set of points of the non-target class}

\item{Osimp}{The indices of the facets associated with the outer simplicies}

\item{x}{The set of reference points of the target class}
}
\value{
A vector indicating the indices of Osimp in which each point of x is found
}
\description{
Search for the outer simplicies that a set of points belong to
}
