% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcdfunctions.R
\name{get_incenter_simp}
\alias{get_incenter_simp}
\title{Get the incenter of an n-simplex}
\usage{
get_incenter_simp(simp)
}
\arguments{
\item{simp}{d+1-by-d matrix representing an d-simplex}
}
\value{
The incenter of the outer simplex
}
\description{
Get the incenter of an n-simplex
}
