% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcdfunctions.R
\name{get_line_par_to}
\alias{get_line_par_to}
\title{Find a parallel ray passing a reference point.}
\usage{
get_line_par_to(x, ray)
}
\arguments{
\item{x}{The Cartesian coordinates of a reference point}

\item{ray}{2-by-d matrix representing a ray.}
}
\value{
A plane passing a reference point x which is parallel to the ray.
}
\description{
Find a parallel ray passing a reference point.
}
