% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcdfunctions.R
\name{get_line_plane_inter}
\alias{get_line_plane_inter}
\title{Find the point of intersection between a ray and a plane.}
\usage{
get_line_plane_inter(ray, plane)
}
\arguments{
\item{ray}{2-by-d matrix representing a ray.}

\item{plane}{d-by-d matrix representing a plane.}
}
\value{
A point on a ray.
}
\description{
Find the point of intersection between a ray and a plane.
}
