% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcdfunctions.R
\name{get_normal_vector}
\alias{get_normal_vector}
\title{Get the normal vector of a plane}
\usage{
get_normal_vector(pla)
}
\arguments{
\item{pla}{N-by-N matrix representing a plane}
}
\value{
The normal vector of the plane.
}
\description{
Get the normal vector of a plane
}
