% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcdfunctions.R
\name{get_plane_par_to}
\alias{get_plane_par_to}
\title{Find a parallel plane passing a reference point.}
\usage{
get_plane_par_to(x, plane)
}
\arguments{
\item{x}{The Cartesian coordinates of a reference point}

\item{plane}{d-by-d matrix representing a plane.}
}
\value{
A plane passing a reference point x which is parallel to the plane.
}
\description{
Find a parallel plane passing a reference point.
}
