% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcdfunctions.R
\name{get_plane_volume}
\alias{get_plane_volume}
\title{Get the volume of a d-simplex.}
\usage{
get_plane_volume(simp)
}
\arguments{
\item{simp}{d+1-by-d matrix representing a simplex.}
}
\value{
The volume of the simplex.
}
\description{
Get the volume of a d-simplex.
}
