% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcdfunctions.R
\name{getDisttoOsimp}
\alias{getDisttoOsimp}
\title{Get the distances of a list of points from an outer simplex.}
\usage{
getDisttoOsimp(Osimp, x)
}
\arguments{
\item{Osimp}{A 2N-by-N matrix representing an outer simplex.}

\item{x}{The set of points of the target class}
}
\value{
The set of distances for each point of x
}
\description{
Get the distances of a list of points from an outer simplex.
}
