% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcdfunctions.R
\name{in_Osimp}
\alias{in_Osimp}
\title{Find if points are in a set of outer simplices}
\usage{
in_Osimp(Osimp, x)
}
\arguments{
\item{Osimp}{A list of d+1-by-d matrices each representing an outer simplex.}

\item{x}{The Cartesian coordinates of a reference point.}
}
\value{
A logical vector indicating in which simplex the reference point x resides.
}
\description{
Find if points are in a set of outer simplices
}
