% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcdfunctions.R
\name{in_simp}
\alias{in_simp}
\title{find if points are in a set of n-simplexes}
\usage{
in_simp(simp, x)
}
\arguments{
\item{simp}{A list of N+1-by-N matrices each representing an n-simplex.}

\item{x}{The Cartesian coordinates of a reference point}
}
\value{
A logical vector indicating in which simplex the reference point x resides.
}
\description{
find if points are in a set of n-simplexes
}
