% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcd-cs.R
\name{pcd_cs_Osimp}
\alias{pcd_cs_Osimp}
\title{Find the pcd of an outer simplex for a given parameter r.}
\usage{
pcd_cs_Osimp(x, y, CM, tau)
}
\arguments{
\item{x}{The set of points of the target class}

\item{y}{The coordinates of the vertices of the outer simplex}

\item{CM}{The median point of the convex hull of the non-target class points}

\item{tau}{The CS-PCD parameter}
}
\value{
The dominating point of the outer simplex and the associated proximity region
}
\description{
Find the pcd of an outer simplex for a given parameter r.
}
