% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcd-pe.R
\name{pcd_pe_simp}
\alias{pcd_pe_simp}
\title{Find the pcd of a single simplex for a given parameter r.}
\usage{
pcd_pe_simp(x, y, bary, r)
}
\arguments{
\item{x}{The set of points of the target class}

\item{y}{The coordinates of the vertices of the n-simplex}

\item{bary}{The barycentric coordinates of the points of x with respect to the n-simplex}

\item{r}{The PE-PCD parameter}
}
\value{
The dominating set of the n-simplex and the associated proximity regions
}
\description{
Find the pcd of a single simplex for a given parameter r.
}
